refactor(api): split module labware handling
This puts all the module labware stuff (module geometries, loading, etc)
into its own file module_geometry.py. This is both a good idea in
general since it reduces the size of the fairly monstrous labware.py and
also paves the way for future work in loading modules.
