+++ api/src/opentrons/commands/commands.py
@@ -1,4 +1,3 @@
-
 from . import types as command_types
 from opentrons.broker import Broker
 
@@ -10,7 +9,8 @@ from opentrons.legacy_api.containers import (Well as OldWell,
                                              Container as OldContainer,
                                              Slot as OldSlot,
                                              location_to_list)
-from opentrons.protocol_api.labware import Well, Labware, ModuleGeometry
+from opentrons.protocol_api.labware import Well, Labware
+from opentrons.protocol_api.module_geometry import ModuleGeometry
 from opentrons.types import Location
 from opentrons.drivers import utils
