+++ api/src/opentrons/protocol_api/definitions.py
@@ -1,4 +1,28 @@
+import abc
 from ..protocols import types
 
-MAX_SUPPORTED_VERSION = types.APIVersion(2, 2)
+MAX_SUPPORTED_VERSION = types.APIVersion(2, 3)
 #: The maximum supported protocol API version in this release
+
+
+class DeckItem(abc.ABC):
+
+    @property  # type: ignore
+    @abc.abstractmethod
+    def highest_z(self):
+        pass
+
+    @highest_z.setter  # type: ignore
+    @abc.abstractmethod
+    def highest_z(self, new_z: float):
+        pass
+
+    @property  # type: ignore
+    @abc.abstractmethod
+    def disambiguate_calibration(self) -> bool:
+        pass
+
+    @property  # type: ignore
+    @abc.abstractmethod
+    def load_name(self) -> str:
+        pass
