+++ api/src/opentrons/protocol_api/labware.py
@@ -1,7 +1,7 @@
 """ opentrons.protocol_api.labware: classes and functions for labware handling
 
-This module provides things like :py:class:`Labware`, :py:class:`Well`, and
-:py:class:`ModuleGeometry` to encapsulate labware instances used in protocols
+This module provides things like :py:class:`Labware`, and :py:class:`Well`
+to encapsulate labware instances used in protocols
 and their wells. It also provides helper functions to load and save labware
 and labware calibration offsets. It contains all the code necessary to
 transform from labware symbolic points (such as "well a1 of an opentrons
@@ -13,24 +13,26 @@ import re
 import time
 import os
 import shutil
-import abc
+
 from pathlib import Path
 from collections import defaultdict
 from enum import Enum, auto
 from hashlib import sha256
 from itertools import takewhile, dropwhile
-from typing import Any, AnyStr, List, Dict, Optional, Union, Sequence, Tuple
+from typing import (
+    Any, AnyStr, List, Dict, Optional, Union, Sequence, Tuple, TYPE_CHECKING)
 
 import jsonschema  # type: ignore
 
-from .util import ModifiedList
+from .util import ModifiedList, requires_version
 from opentrons.types import Location, Point
 from opentrons.config import CONFIG
 from opentrons.protocols.types import APIVersion
 from opentrons.system.shared_data import load_shared_data, get_shared_data_root
-from .definitions import MAX_SUPPORTED_VERSION
+from .definitions import MAX_SUPPORTED_VERSION, DeckItem
+if TYPE_CHECKING:
+    from .module_geometry import ModuleGeometry  # noqa(F401)
 
-from .util import requires_version
 
 MODULE_LOG = logging.getLogger(__name__)
 
@@ -56,19 +58,6 @@ well_shapes = {
     'rectangular': WellShape.RECTANGULAR,
     'circular': WellShape.CIRCULAR
 }
-
-
-class DeckItem(abc.ABC):
-
-    @property  # type: ignore
-    @abc.abstractmethod
-    def highest_z(self):
-        pass
-
-    @highest_z.setter  # type: ignore
-    @abc.abstractmethod
-    def highest_z(self, new_z: float):
-        pass
 
 
 class Well:
@@ -337,6 +326,10 @@ class Labware(DeckItem):
         self._pattern = re.compile(r'^([A-Z]+)([1-9][0-9]*)$', re.X)
         self._definition = definition
         self._highest_z = self._dimensions['zDimension']
+
+    @property
+    def disambiguate_calibration(self) -> bool:
+        return False
 
     @property  # type: ignore
     @requires_version(2, 0)
@@ -806,185 +799,9 @@ class Labware(DeckItem):
                 well.has_tip = True
 
 
-class ModuleGeometry(DeckItem):
-    """
-    This class represents an active peripheral, such as an Opentrons Magnetic
-    Module, Temperature Module or Thermocycler Module. It defines the physical
-    geometry of the device (primarily the offset that modifies the position of
-    the labware mounted on top of it).
-    """
-
-    @classmethod
-    def resolve_module_name(cls, module_name: str):
-        alias_map = {
-            'magdeck': 'magdeck',
-            'magnetic module': 'magdeck',
-            'tempdeck': 'tempdeck',
-            'temperature module': 'tempdeck',
-            'thermocycler': 'thermocycler',
-            'thermocycler module': 'thermocycler'
-        }
-        lower_name = module_name.lower()
-        resolved_name = alias_map.get(lower_name, None)
-        if not resolved_name:
-            raise ValueError(f'{module_name} is not a valid module load name')
-        return resolved_name
-
-    def __init__(self,
-                 definition: dict,
-                 parent: Location,
-                 api_level: APIVersion = None) -> None:
-        """
-        Create a Module for tracking the position of a module.
-
-        Note that modules do not currently have a concept of calibration apart
-        from calibration of labware on top of the module. The practical result
-        of this is that if the module parent :py:class:`.Location` is
-        incorrect, then acorrect calibration of one labware on the deck would
-        be incorrect on the module, and vice-versa. Currently, the way around
-        this would be to correct the :py:class:`.Location` so that the
-        calibrated labware is targeted accurately in both positions.
-
-        :param definition: A dict containing all the data required to define
-                           the geometry of the module.
-        :type definition: dict
-        :param parent: A location representing location of the front left of
-                       the outside of the module (usually the front-left corner
-                       of a slot on the deck).
-        :type parent: :py:class:`.Location`
-        :param APIVersion api_level: the API version to set for the loaded
-                                     :py:class:`ModuleGeometry` instance. The
-                                     :py:class:`ModuleGeometry` will
-                                     conform to this level. If not specified,
-                                     defaults to
-                                     :py:attr:`.MAX_SUPPORTED_VERSION`.
-        """
-        if not api_level:
-            api_level = MAX_SUPPORTED_VERSION
-        if api_level > MAX_SUPPORTED_VERSION:
-            raise RuntimeError(
-                f'API version {api_level} is not supported by this '
-                f'robot software. Please either reduce your requested API '
-                f'version or update your robot.')
-        self._api_version = api_level
-        self._parent = parent
-        self._display_name = "{} on {}".format(definition["displayName"],
-                                               str(parent.labware))
-        self._load_name = definition["loadName"]
-        self._offset = Point(definition["labwareOffset"]["x"],
-                             definition["labwareOffset"]["y"],
-                             definition["labwareOffset"]["z"])
-        self._height = definition["dimensions"]["bareOverallHeight"]\
-            + self._parent.point.z
-        self._over_labware = definition["dimensions"]["overLabwareHeight"]
-        self._labware: Optional[Labware] = None
-        self._location = Location(
-            point=self._offset + self._parent.point,
-            labware=self)
-
-    @property
-    def api_version(self) -> APIVersion:
-        return self._api_version
-
-    def add_labware(self, labware: Labware) -> Labware:
-        assert not self._labware,\
-            '{} is already on this module'.format(self._labware)
-        self._labware = labware
-        return self._labware
-
-    def reset_labware(self):
-        self._labware = None
-
-    @property
-    def load_name(self):
-        return self._load_name
-
-    @property
-    def parent(self):
-        return self._parent.labware
-
-    @property
-    def labware(self) -> Optional[Labware]:
-        return self._labware
-
-    @property
-    def location(self) -> Location:
-        """
-        :return: a :py:class:`.Location` representing the top of the module
-        """
-        return self._location
-
-    @property
-    def labware_offset(self) -> Point:
-        """
-        :return: a :py:class:`.Point` representing the transformation
-        between the critical point of the module and the critical
-        point of its contained labware
-        """
-        return self._offset
-
-    @property
-    def highest_z(self) -> float:
-        if self.labware:
-            return self.labware.highest_z + self._over_labware
-        else:
-            return self._height
-
-    def __repr__(self):
-        return self._display_name
-
-
-class ThermocyclerGeometry(ModuleGeometry):
-    def __init__(self, definition: Dict[str, Any], parent: Location,
-                 api_level: APIVersion = None) -> None:
-        super().__init__(definition, parent, api_level)
-        self._lid_height = definition["dimensions"]["lidHeight"]
-        self._lid_status = 'open'   # Needs to reflect true status
-        # TODO: BC 2019-07-25 add affordance for "semi" configuration offset
-        # to be from a flag in context, according to drawings, the only
-        # difference is -23.28mm in the x-axis
-
-    @property
-    def highest_z(self) -> float:
-        # TODO: BC 2019-08-27 this highest_z value represents the distance
-        # from the top of the open TC chassis to the base. Once we have a
-        # more robust collision detection system in place, the collision
-        # model for the TC should change based on it's lid_status
-        # (open or closed). A prerequisite for that check will be
-        # path-specific highest z calculations, as opposed to the current
-        # global check on instrument.move_to. For example: a move from slot 1
-        # to slot 3 should only check the highest z of all deck items between
-        # the source and destination in the x,y plane.
-        return super().highest_z
-
-    @property
-    def lid_status(self) -> str:
-        return self._lid_status
-
-    @lid_status.setter
-    def lid_status(self, status) -> None:
-        self._lid_status = status
-
-    # NOTE: this func is unused until "semi" configuration
-    def labware_accessor(self, labware: Labware) -> Labware:
-        # Block first three columns from being accessed
-        definition = labware._definition
-        definition['ordering'] = definition['ordering'][3::]
-        return Labware(
-            definition, super().location, api_level=self._api_version)
-
-    def add_labware(self, labware: Labware) -> Labware:
-        assert not self._labware,\
-            '{} is already on this module'.format(self._labware)
-        assert self.lid_status != 'closed', \
-            'Cannot place labware in closed module'
-        self._labware = labware
-        return self._labware
-
-
 def _get_parent_identifier(
-        parent: Union[Labware, Well, str, ModuleGeometry, None]):
-    if isinstance(parent, ModuleGeometry):
+        parent: Union[Well, str, DeckItem, None]):
+    if isinstance(parent, DeckItem) and parent.disambiguate_calibration:
         # treat a given labware on a given module type as same
         return parent.load_name
     else:
@@ -1341,62 +1158,8 @@ def clear_calibrations():
         pass
 
 
-def load_module_from_definition(
-        definition: Dict[str, Any],
-        parent: Location,
-        api_level: APIVersion = None) -> ModuleGeometry:
-    """
-    Return a :py:class:`ModuleGeometry` object from a specified definition
-
-    :param definition: A dict representing all required data for a module's
-                       geometry.
-    :param parent: A :py:class:`.Location` representing the location where
-                   the front and left most point of the outside of the module
-                   is (often the front-left corner of a slot on the deck).
-    :param APIVersion api_level: the API version to set for the loaded
-                                 :py:class:`ModuleGeometry` instance. The
-                                 :py:class:`ModuleGeometry` will
-                                 conform to this level. If not specified,
-                                 defaults to :py:attr:`.MAX_SUPPORTED_VERSION`.
-    """
-    mod_name = definition['loadName']
-    if mod_name == 'thermocycler':
-        mod: ModuleGeometry = \
-                ThermocyclerGeometry(definition, parent, api_level)
-    else:
-        mod = ModuleGeometry(definition, parent, api_level)
-    # TODO: calibration
-    return mod
-
-
-def load_module(
-        name: str,
-        parent: Location,
-        api_level: APIVersion = None) -> ModuleGeometry:
-    """
-    Return a :py:class:`ModuleGeometry` object from a definition looked up
-    by name.
-
-    :param name: A string to use for looking up the definition. The string
-                 must be present as a top-level key in
-                 module/definitions/{moduleDefinitionVersion}.json.
-    :param parent: A :py:class:`.Location` representing the location where
-                   the front and left most point of the outside of the module
-                   is (often the front-left corner of a slot on the deck).
-    :param APIVersion api_level: the API version to set for the loaded
-                                 :py:class:`ModuleGeometry` instance. The
-                                 :py:class:`ModuleGeometry` will
-                                 conform to this level. If not specified,
-                                 defaults to :py:attr:`.MAX_SUPPORTED_VERSION`.
-    """
-    def_path = 'module/definitions/1.json'
-    module_def = json.loads(
-        load_shared_data(def_path).decode('utf-8'))  # type: ignore
-    return load_module_from_definition(module_def[name], parent, api_level)
-
-
 def quirks_from_any_parent(
-        loc: Union[Labware, Well, str, ModuleGeometry, None]) -> List[str]:
+        loc: Union[Labware, Well, str, 'ModuleGeometry', None]) -> List[str]:
     """ Walk the tree of wells and labwares and extract quirks """
     def recursive_get_quirks(obj, found):
         if isinstance(obj, Labware):
