+++ api/src/opentrons/types.py
@@ -5,7 +5,8 @@ if TYPE_CHECKING:
     from typing import (Optional,       # noqa(F401) Used for typechecking
                         Tuple)
     from .protocol_api.labware import (  # noqa(F401) Used for typechecking
-        Labware, Well, ModuleGeometry)
+        Labware, Well)
+    from .protocol_api.module_geometry import ModuleGeometry  # noqa(F401)
 
 
 class PipetteNotAttachedError(KeyError):
